+++ src/Autocomplete/Autocomplete.stories.tsx
@@ -1,5 +1,5 @@
 /* ======================================================================== *
- * Copyright 2024, 2025 HCL America Inc.                                    *
+ * Copyright 2026 HCL America Inc.                                          *
  * Licensed under the Apache License, Version 2.0 (the "License");          *
  * you may not use this file except in compliance with the License.         *
  * You may obtain a copy of the License at                                  *
@@ -16,6 +16,8 @@
 import React from 'react';
 import { StoryFn, Meta } from '@storybook/react';
 import { userEvent, within } from '@storybook/testing-library';
+import CaretDownIcon from '@hcl-software/enchanted-icons/dist/carbon/es/caret--down';
+import InformationIcon from '@hcl-software/enchanted-icons/dist/carbon/es/information';
 
 import Autocomplete from './Autocomplete';
 import { top100Films } from './data';
@@ -210,6 +212,16 @@ export default {
       },
       description: 'https://mui.com/material-ui/api/text-field/',
     },
+    customIcon: {
+      description: 'This can be used to add a custom icon replacing the default information icon for helper text.',
+      options: ['None', 'CaretDownIcon', 'InformationIcon'],
+      control: { type: 'radio' },
+      table: {
+        defaultValue: {
+          summary: 'None',
+        },
+      },
+    },
   },
 } as Meta<typeof Autocomplete>;
 
@@ -220,6 +232,19 @@ interface Movie {
 
 const Template: StoryFn<typeof Autocomplete> = (args) => {
   const [value, setValue] = React.useState(args.value ? args.value : null);
+
+  let customIcon: React.ComponentType<React.SVGProps<SVGSVGElement>> | undefined;
+  switch (args.customIcon as unknown as string) {
+    case 'CaretDownIcon':
+      customIcon = CaretDownIcon;
+      break;
+    case 'InformationIcon':
+      customIcon = InformationIcon;
+      break;
+    default:
+      customIcon = undefined;
+  }
+
   return (
     <Autocomplete
       value={value}
@@ -246,6 +271,7 @@ const Template: StoryFn<typeof Autocomplete> = (args) => {
         );
       }}
       {...args}
+      customIcon={customIcon}
     />
   );
 };
@@ -290,6 +316,7 @@ export const ExampleAutocomplete = {
     ],
     options: top100Films,
     sx: { minWidth: '240px' },
+    customIcon: 'None',
   },
 };